+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 * Account for distance between radiation emitters (NERVs, RTGs etc.) and habitats when calculating the effect of radiation (Sir Mortimer)
 * EVAs will receive radiation from nearby emitters (Sir Mortimer)
 * Lead was replaced by a lighter alloy composite as a shielding material. For our type of radiation, it provides the same shielding effect with less mass. (Sir Mortimer)
+* Warn users about incompatible mods or missing CommunityResourcePack (Sir Mortimer)
 
 ## v3.0.2 for all versions of KSP from 1.4.0 to 1.7.x